### 2023-11-21 #6075 @danielskatz
👋 @rcannood - JOSS works best when the paper branch contains the software as well, not just the paper. Can you please change the paper branch of be a copy of main that also includes the paper, rather than just the paper?
