### 2023-11-21 #6075 @danielskatz
Also, your figures are not appearing the draft paper. Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). Please feel free to make changes to your .md file, then use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.
