+++ docs/paper/paper.md
@@ -1,0 +1,120 @@
+---
+title: 'Viash: A meta-framework for building reusable workflow modules'
+bibliography: library.bib
+authors:
+- name: Robrecht Cannoodt
+  email: robrecht@data-intuitive.com
+  orcid: 0000-0003-3641-729X
+  corresponding: yes
+  affiliation: 1, 2, 3
+- name: Hendrik Cannoodt
+  affiliation: '1'
+- name: Dries Schaumont
+  affiliation: '1'
+- name: Kai Waldrant
+  affiliation: '1'
+- name: Eric Van de Kerckhove
+  affiliation: '1'
+- name: Andy Boschmans
+  affiliation: '1'
+- name: Dries De Maeyer
+  affiliation: '4'
+- name: Toni Verbeiren
+  email: toni@data-intuitive.com
+  corresponding: yes
+  affiliation: '1'
+affiliations:
+- name: Data Intuitive, Lebbeke, Belgium
+  index: 1
+- name: Data Mining and Modelling for Biomedicine group, VIB Center for Inflammation
+    Research, Ghent, Belgium
+  index: 2
+- name: Department of Applied Mathematics, Computer Science, and Statistics, Ghent
+    University, Ghent, Belgium
+  index: 3
+- name: Discovery Technology and Molecular Pharmacology, Janssen Research & Development,
+    Pharmaceutical Companies of Johnson & Johnson, Beerse, Belgium
+  index: 4
+tags:
+- Bioinformatics
+- Workflows
+- Software development
+- Reproducibility
+- Cloud computing
+- Docker
+- Nextflow
+---
+
+
+
+
+
+# Abstract
+Most bioinformatics pipelines consist of software components that are tightly coupled to the logic of the pipeline itself. This limits reusability of the individual components in the pipeline or introduces maintenance overhead when they need to be reimplemented in multiple pipelines. We introduce Viash, a tool for speeding up development of robust pipelines through "code-first" prototyping, separation of concerns and code generation of modular pipeline components. By decoupling the component functionality from the pipeline logic, component functionality becomes fully pipeline-agnostic, and conversely the resulting pipelines are agnostic towards specific component requirements. This separation of concerns improves reusability of components and facilitates multidisciplinary and pan-organisational collaborations. It has been applied in a variety of projects, from proof-of-concept pipelines to supporting an international data science competition.
+
+Viash is available as an open-source project at [github.com/viash-io/viash](https://github.com/viash-io/viash) and documentation is available at [viash.io](https://viash.io).
+
+
+
+# Statement of Need
+Recent developments in high-throughput RNA sequencing and imaging technologies allow present-day biologists to observe single-cell characteristics in ever more detail [@luecken_currentbestpractices_2019]. As the dataset size and the complexity of bioinformatics workflows increases, so does the need for scalable and reproducible data science. In single cell biology, recent efforts to standardise some of the most common single-cell analyses [@amezquita_orchestratingsinglecell_2020; @nfcoreframeworkcommunity_ewels2020; @huemos_bestpracticessingle_2023] tackle these challenges by using a pipeline frameworks (e.g. Snakemake [@koster_snakemakescalablebioinformaticsworkflow_2012], Nextflow [@ditommaso_nextflowenablesreproducible_2017]), containerisation (e.g. Docker [@merkel_dockerlightweight_2014], Singularity/Apptainer [@kurtzer_singularityscientificcontainers_2017], Podman [@heon_podmantoolmanaging_2018]) and horizontal scaling in cloud computing (e.g. Kubernetes, HPC).
+
[…]
