+++ docs/paper/library.bib
@@ -1,0 +1,54 @@
+ @article{koster_snakemakescalablebioinformaticsworkflow_2012, title={Snakemake—a scalable bioinformatics workflow engine}, volume={28}, ISSN={1367-4803}, url={http://dx.doi.org/10.1093/bioinformatics/bts480}, DOI={10.1093/bioinformatics/bts480}, number={19}, journal={Bioinformatics}, publisher={Oxford University Press (OUP)}, author={Köster, Johannes and Rahmann, Sven}, year={2012}, month=aug, pages={2520–2522} }
+
+ @article{amezquita_orchestratingsinglecell_2020, title={Orchestrating single-cell analysis with Bioconductor}, volume={17}, ISSN={1548-7105}, url={http://dx.doi.org/10.1038/s41592-019-0654-x}, DOI={10.1038/s41592-019-0654-x}, number={2}, journal={Nature Methods}, publisher={Springer Science and Business Media LLC}, author={Amezquita, Robert A. and Lun, Aaron T. L. and Becht, Etienne and Carey, Vince J. and Carpp, Lindsay N. and Geistlinger, Ludwig and Marini, Federico and Rue-Albrecht, Kevin and Risso, Davide and Soneson, Charlotte and Waldron, Levi and Pagès, Hervé and Smith, Mike L. and Huber, Wolfgang and Morgan, Martin and Gottardo, Raphael and Hicks, Stephanie C.}, year={2019}, month=dec, pages={137–145} }
+
+ @article{huemos_bestpracticessingle_2023, title={Best practices for single-cell analysis across modalities}, volume={24}, ISSN={1471-0064}, url={http://dx.doi.org/10.1038/s41576-023-00586-w}, DOI={10.1038/s41576-023-00586-w}, number={8}, journal={Nature Reviews Genetics}, publisher={Springer Science and Business Media LLC}, author={Heumos, Lukas and Schaar, Anna C. and Lance, Christopher and Litinetskaya, Anastasia and Drost, Felix and Zappia, Luke and Lücken, Malte D. and Strobl, Daniel C. and Henao, Juan and Curion, Fabiola and Aliee, Hananeh and Ansari, Meshal and Badia-i-Mompel, Pau and Büttner, Maren and Dann, Emma and Dimitrov, Daniel and Dony, Leander and Frishberg, Amit and He, Dongze and Hediyeh-zadeh, Soroor and Hetzel, Leon and Ibarra, Ignacio L. and Jones, Matthew G. and Lotfollahi, Mohammad and Martens, Laura D. and Müller, Christian L. and Nitzan, Mor and Ostner, Johannes and Palla, Giovanni and Patro, Rob and Piran, Zoe and Ramírez-Suástegui, Ciro and Saez-Rodriguez, Julio and Sarkar, Hirak and Schubert, Benjamin and Sikkema, Lisa and Srivastava, Avi and Tanevski, Jovan and Virshup, Isaac and Weiler, Philipp and Schiller, Herbert B. and Theis, Fabian J.}, year={2023}, month=mar, pages={550–572} }
+
+ @article{luecken_currentbestpractices_2019, title={Current best practices in single‐cell RNA‐seq analysis: a tutorial}, volume={15}, ISSN={1744-4292}, url={http://dx.doi.org/10.15252/msb.20188746}, DOI={10.15252/msb.20188746}, number={6}, journal={Molecular Systems Biology}, publisher={EMBO}, author={Luecken, Malte D and Theis, Fabian J}, year={2019}, month=jun }
+
[…]
