### 2023-11-22 #6075 @rcannood
Thanks @danielskatz ! I'll make some changes to make the paper render properly. It's strange because when I rendered the paper locally using `docker run ... openjournals/inara`, everything looks ok.
