### 2023-11-22 #6075 @rcannood
@danielskatz Here is a potential list of reviewers:

* mberacochea - Previously reviewed [Baargin: a Nextflow workflow ...](https://joss.theoj.org/papers/10.21105/joss.05397) and [ALPPACA](https://joss.theoj.org/papers/10.21105/joss.04677) together with me. Has a lot of activity related to Nextflow workflows and modules in his GitHub Contribution activity.
* hseabolt - Also reviewed ALPPACA together with me, and has activity related to Nextflow workflows and modules in his GitHub Contribution activity.
* Zethson - Previously reviewed [BioProv - A provenance library for bioinformatics workflows](https://github.com/openjournals/joss-reviews/issues/3622), and also has activity related to Nextflow workflows in his GitHub Contribution activity
