### 2023-11-22 #6075 @danielskatz
@rcannood - thanks for your changes. I'll assign myself as the editor and get started on looking for reviewers.

As a note, this paper is fairly long now (JOSS recommends 1000 words as a maximum), so if you add anything during the review process, please find something else to remove at the same time.
