### 2023-11-22 #6075 @danielskatz
👋 @mberacochea, @hseabolt, @Zethson - would any of you be willing to review this JOSS submission?
