### 2023-11-23 #6075 @abhishektiwari
@danielskatz I am open to review workflow related submissions. Feel free to assign me as additional reviewer on this. Seems like an intersting approach to build  modular data pipelines.
