### 2023-11-24 #6089 @editorialbot
[REVIEW]: Viash: A meta-framework for building reusable workflow modules
**Submitting author:** @rcannood (<a href="http://orcid.org/0000-0003-3641-729X">Robrecht Cannoodt</a>)
**Repository:** https://github.com/viash-io/viash
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.8.4
