+++ .github/workflows/sbt_test.yml
@@ -43,7 +43,7 @@ jobs:
           testthat
 
     - name: Set up sbt
-      uses: actions/setup-java@v3
+      uses: actions/setup-java@v4
       with:
         distribution: temurin
         java-version: ${{ matrix.java.ver }}
