+++ .github/workflows/ns_test.yml
@@ -10,7 +10,7 @@ jobs:
     - uses: actions/checkout@v4
 
     - name: Set up sbt
-      uses: actions/setup-java@v3
+      uses: actions/setup-java@v4
       with:
         distribution: temurin
         java-version: '11'
