### 2023-12-01 acff5fc
use Java 21 LTS instead of Java 20
Co-authored-by: Hendrik Cannoodt <hendrik.cannoodt@gmail.com>
