### 2023-12-01 0797882
Revert "use Java 21 LTS instead of Java 20"
This reverts commit acff5fc34dc36e39ce08e128fb7ff5f80cff4c49.
