### 2023-12-01 0b4cade
Test different versions of nextflow (#593)
* test different versions of nextflow

* update changelog

* update setup-java

* use Java 21 LTS instead of Java 20

Co-authored-by: Hendrik Cannoodt <hendrik.cannoodt@gmail.com>

* Revert "use Java 21 LTS instead of Java 20"

This reverts commit acff5fc34dc36e39ce08e128fb7ff5f80cff4c49.

---------

Co-authored-by: Hendrik Cannoodt <hendrik.cannoodt@gmail.com>
