### 2023-12-01 732bd80
Add function tests
