### 2023-12-01 5ce0d31
Run tests only when nextflow installed
