### 2023-12-01 3dfcfe7
Revert CHANGELOG changes
