+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-# Viash 0.8.2 (yyyy-MM-dd): TODO Add title
+# Viash 0.x.x (yyyy-MM-dd): TODO Add title
+
+TODO add summary
 
 ## BUG FIXES
 
