### 2023-12-04 #6089 @danielskatz
👋 @wilkinson and @mberacochea - I would appreciate it if you could use the command `@editorialbot generate my checklist` to create your review checklist, and then check off the CoI and CoC items, to make everything works. @editorialbot commands need to be the first thing in a new comment.
