+++ .github/workflows/sbt_test.yml
@@ -61,7 +61,7 @@ jobs:
         fi
 
     - name: Set up Python
-      uses: actions/setup-python@v4
+      uses: actions/setup-python@v5
       with:
         python-version: '3.x'
 
