### 2023-12-10 viash-io/viash#598 @abhishektiwari
Add Community guidelines
Add clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

For [JOSS Review] (https://github.com/openjournals/joss-reviews/issues/6089)
