### 2023-12-10 #6089 @abhishektiwari
@rcannood Please see my initial feedback.

# Documentation

## Community Guidelines
Please add [community guidelines](https://github.com/viash-io/viash/issues/598). I also think project will benefit from issue and pull request templates.

# Paper

## Title
For consistency, I would suggest paper headline to use `data pipeline modules` and not `workflow modules`.  Everywhere else in the paper and documentation website authors reference Viash as solution for modular `data pipelines`. Having said that data pipelines are a special type of workflows with major focus on ETL, data cleaning, integration, and preparation for analysis or visualization, so I am also fine using `workflow`.

## State of the field
Current state of the field can be further improved. Given the focus on Bioinformatics related computation/data pipeline frameworks, Viash, and it's counterparts Nextflow, Snakemake, CWL belong to `Pipeline as code` category.  But there is other category of frameworks such as Galaxy and KNIME Workbench used by bioinformatics community those rely on `UI based drag-n-drop` pipeline builders – something worth mentioning here. Similarly, more generic `Pipeline as code` frameworks such Apache Airflow which can use the CLI output generated by the Viash.

# Suggested improvements (non-blocking)
Given the Viash has broader applicability beyond Bioinformatics particularly MLOps Pipelines, I suggest adding starter using other more generic frameworks such Airflow.  In case of Apache Airflow, Airflow [BashOperator](https://airflow.apache.org/docs/apache-airflow/stable/howto/operator/bash.html) will works just fine with standalone executable generated by Viash.

# Overall comments
I applaud the author's authors for providing a very frictionless experience. Both Guide and Reference are well documented. I was able to finish everything from installation to functional testing to example in less than an hour. I really liked the use of GitHub template to create a quickstarter. Furthermore, I was able to add new modules on top of existing quickstarter very quickly and run end-to-end pipeline. Finally, it was great to see that authors provided guidance on IDE setup and testing of modules created using Viash.
