### 2023-12-11 fa3c3ca
Update CONTRIBUTING.md
Co-authored-by: Hendrik Cannoodt <hendrik.cannoodt@gmail.com>
