### 2023-12-11 522b476
add staging command
Co-authored-by: Hendrik Cannoodt <hendrik.cannoodt@gmail.com>
