### 2023-12-11 1fbd1b5
uniformize contributing guidelines
