### 2023-12-11 1655500
Bump scala to 2.13.12 and CI java version from 20 to 21
Attempt to run Nextflow with Java 21 too.
