### 2023-12-11 1e3db7d
add contributing guidelines (#600)
* add contributing guidelines

* update changelog

* add more templates

* Update CONTRIBUTING.md

Co-authored-by: Hendrik Cannoodt <hendrik.cannoodt@gmail.com>

* add staging command

Co-authored-by: Hendrik Cannoodt <hendrik.cannoodt@gmail.com>

* uniformize contributing guidelines

* update contributing info

---------

Co-authored-by: Hendrik Cannoodt <hendrik.cannoodt@gmail.com>
