+++ .github/pull_request_template.md
@@ -1,18 +1,45 @@
 ## Describe your changes
 
-## Issue ticket number and link
-Closes #xxxx (Replace xxxx with the GitHub issue number)
+<!-- 
+Please include a summary of the change and which issue is fixed.
+-->
 
-## Checklist before requesting a review
-- [ ] I have performed a self-review of my code
+## Related issue(s)
 
-- Does this PR contain:
-  - [ ] Breaking changes
-  - [ ] New functionality
-  - [ ] Major changes
-  - [ ] Minor changes
-  - [ ] Bug fixes
+<!-- Replace xxxx with the GitHub issue number. -->
 
-- [ ] Proposed changes are described in the CHANGELOG.md
+Closes #xxxx 
 
-- [ ] Relevant unit tests have been added
+## Type of Change
+
+- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
+- [ ] New functionality (non-breaking change which adds functionality)
+- [ ] Major change (non-breaking change which modifies existing functionality)
+- [ ] Minor change (non-breaking change which does not modify existing functionality)
+- [ ] Bug fix (non-breaking change which fixes an issue)
+- [ ] This change requires a documentation update
+
+## Checklist
+
+Requirements:
+
+- [ ] I have read the [CONTRIBUTING](CONTRIBUTING.md) doc.
+- [ ] I have performed a self-review of my code by checking the "Changed Files" tab.
+- [ ] My code follows the code style of this project.
+
+Tests:
+
+- [ ] I have added tests that prove my fix is effective or that my feature works.
+- [ ] New and existing unit tests pass locally with my changes.
+
+Documentation:
+
+- [ ] Proposed changes are described in the CHANGELOG.md.
+- [ ] I have updated the documentation accordingly.
+
+## Test Environment
+
+<!--
+Please describe the tests that you ran to verify your changes. Provide instructions so we can reproduce. Please also list any relevant details for your test environment.
+-->
+
