### 2023-12-11 893be76
disable nextflow tests again with Java version 21
Java 21 support was added to Nextflow in version 23.10.0
NextflowTestHelper.scala forces Nextflow to version 22.04.5
