### 2023-12-11 fcacdef
Add a changelog entry
