### 2023-12-11 76c37ae
Merge pull request #585 from viash-io/fix/non_strict_config
Fix config decoding not being enforcing strict mode
