### 2023-12-12 4225126
Merge arguments into argument_groups in preparse step (#574)
* Merge arguments into argument_groups in preparse step

Fixes #573

* Add changelog entry

* Add unit tests covering basic possibilities of arguments and argument_groups

* simplify ArgumentGroup verification

* Arguments should not be outputted in a strict schema

* Throw away the warning when an argument group is specified with strings

The functionality was removed in Viash 0.7.0

* Fix severely outdated comment

* Use automatic type deriving again for argument groups

* Remove test that is no longer valid as the code it tests wa
[…]
