+++ src/main/scala/io/viash/functionality/package.scala
@@ -22,6 +22,7 @@ import io.circe.{Decoder, Encoder, Json}
 import io.circe.ACursor
 
 import io.viash.helpers.Logging
+import io.circe.JsonObject
 
 package object functionality extends Logging {
   // import implicits
@@ -39,8 +40,69 @@ package object functionality extends Logging {
   // encoder and decoder for Functionality
   implicit val encodeFunctionality: Encoder.AsObject[Functionality] = deriveConfiguredEncoderStrict[Functionality]
 
-  // add file & direction defaults for inputs & outputs
-  implicit val decodeFunctionality: Decoder[Functionality] = deriveConfiguredDecoderFullChecks
+  // merge arguments and argument_groups into argument_groups
+  implicit val decodeFunctionality: Decoder[Functionality] = deriveConfiguredDecoderFullChecks[Functionality]
+    .prepare {
+      _.withFocus{ json =>
+        json.asObject match {
+          case None => json
+          case Some(jo) => 
+            val arguments = jo.apply("arguments")
+            val argument_groups = jo.apply("argument_groups")
+
+            val newJsonObject = (arguments, argument_groups) match {
+              case (None, _) => jo
+              case (Some(args), None) => 
+                jo.add("argument_groups", Json.fromValues(List(
+                  Json.fromJsonObject(
+                    JsonObject(
+                      "name" -> Json.fromString("Arguments"),
+                      "arguments" -> args
+                    )
+                  )
+                )))
+              case (Some(args), Some(arg_groups)) =>
+                // determine if we should prepend or append arguments to argument_groups
+                val prepend = jo.keys.find(s => s == "argument_groups" || s == "arguments") == Some("arguments")
+                def combinerSeq(a: Vector[Json], b: Seq[Json]) = if (prepend) b ++: a else a :++ b
+                def combiner(a: Vector[Json], b: Json) = if (prepend) b +: a else a :+ b
+
+                // get the argument group named 'Arguments' from arg_groups
+                val argumentsGroup = arg_groups.asArray.flatMap(_.find(_.asObject.exists(_.apply("name").exists(_ == Json.fromString("Arguments")))))
+                argumentsGroup match {
+                  case None =>
+                    // no argument_group with name 'Argument' exists, so just add arguments as a new argument group
+                    jo.add("argument_groups", Json.fromValues(
+                      combiner(
+                        arg_groups.asArray.get,
+                        Json.fromJsonObject(
+                          JsonObject(
+                            "name" -> Json.fromString("Arguments"),
+                            "arguments" -> args
+                          )
+                        )
+                      )
+                    ))
+                  case Some(ag) =>
+                    // argument_group with name 'Argument' exists, so add arguments to this argument group
+                    val newAg = ag.asObject.get.add("arguments",
+                      Json.fromValues(
+                        combinerSeq(
+                          ag.asObject.get.apply("arguments").get.asArray.get,
+                          args.asArray.get
+                        )
+                      )
+                    )
+                    jo.add("argument_groups", Json.fromValues(arg_groups.asArray.get.map{
+                      case ag if ag == argumentsGroup.get => Json.fromJsonObject(newAg)
+                      case ag => ag
+                    }))
+                }
+            }
+            Json.fromJsonObject(newJsonObject.remove("arguments"))
+        }
+      
+    }}
 
   // encoder and decoder for Author
   implicit val encodeAuthor: Encoder.AsObject[Author] = deriveConfiguredEncoder
@@ -52,37 +114,7 @@ package object functionality extends Logging {
   
   // encoder and decoder for ArgumentGroup
[…]
