### 2023-12-12 ef04c15
Feature/bump scala java versions (#602)
* Bump scala to 2.13.12 and CI java version from 20 to 21

Attempt to run Nextflow with Java 21 too.

* Bump the sbt-scroverage version to one that's available for Scala 2.13.12

* Bump sbt to latest version too

* disable nextflow tests again with Java version 21

Java 21 support was added to Nextflow in version 23.10.0
NextflowTestHelper.scala forces Nextflow to version 22.04.5

* Add a changelog entry
