### 2023-12-12 8ace788
move groovy tests to be run by scala
