### 2023-12-12 cdce592
normalize path in nxf test
