### 2023-12-12 afc7661
Don't require name to be filled in for repositories defined under dependencies
Duplicate repository types to with and without the `name` field.
Define a series of traits that connects the repositories with and without names
