### 2023-12-12 949789d
Fix some generalizations between repository traits and specialized constructor
