### 2023-12-12 dabfb3b
Add changelog entry
Update comments; we no longer have to anonymize repositories.
