### 2023-12-12 4cc982a
Remove copyRepoWithName as it isn't used
