### 2023-12-13 525b6df
use "local" to match the repository name instead of using that in the sugar syntax regex
regex was needed to extract repo and tag names, but this is irrelevant here. It would have required something like `local://...` notation.
