+++ src/main/scala/io/viash/functionality/dependencies/Repository.scala
@@ -83,7 +83,7 @@ object Repository {
           repo = repo,
           tag = getGitTag(tag)
         ))
-      case sugarSyntaxRegex("local", repo, tag) =>
+      case "local" =>
         Some(LocalRepository("TODO generate name"))
       case _ => None
     }
