### 2023-12-13 1609e75
Add sugar syntax notation for local repositories, but now properly
