### 2023-12-13 f3d98bb
Add test for the new sugar syntax option
