### 2023-12-13 6a3d96d
Fix/dependency repositories without name (#607)
* Don't require name to be filled in for repositories defined under dependencies

Duplicate repository types to with and without the `name` field.
Define a series of traits that connects the repositories with and without names

* Fix some generalizations between repository traits and specialized constructor

* Add changelog entry

Update comments; we no longer have to anonymize repositories.

* Remove copyRepoWithName as it isn't used
