+++ src/main/scala/io/viash/functionality/dependencies/RepositoryWithName.scala
@@ -17,13 +17,15 @@
 
 package io.viash.functionality.dependencies
 
-trait CopyableRepo[A <: CopyableRepo[A]] {
-  self: A =>
-  def copyRepo(
-   `type`: String,
-    tag: Option[String],
-    path: Option[String],
-    localPath: String
-  ): A
-}
+import io.viash.schemas._
 
+@description("Specifies a repository where dependency components can be found.")
+@subclass("LocalRepositoryWithName")
+@subclass("GitRepositoryWithName")
+@subclass("GithubRepositoryWithName")
+@subclass("ViashhubRepositoryWithName")
+abstract class RepositoryWithName extends Repository {
+  @description("The identifier used to refer to this repository from dependencies.")
+  val name: String
+
+}
