### 2023-12-13 9a13dd8
use "local" to match the repository name instead of using that in the sugar syntax regex (#609)
* use "local" to match the repository name instead of using that in the sugar syntax regex

regex was needed to extract repo and tag names, but this is irrelevant here. It would have required something like `local://...` notation.

* Add changelog entry

* Add sugar syntax notation for local repositories, but now properly

* Add test for the new sugar syntax option
