### 2023-12-13 fdac38c
Define some extra fields in Viash Project Config and add it to a component info field
also adding an array of strings in a component info field to store cli config mods
Fix some warnings about inference of implicit functions
