+++ src/main/scala/io/viash/config/Info.scala
@@ -18,6 +18,7 @@
 package io.viash.config
 
 import io.viash.schemas.description
+import io.viash.project.ViashProject
 
 // TODO: rename to ConfigInfo?
 
@@ -40,5 +41,10 @@ case class Info(
   @description("Git remote name.")
   git_remote: Option[String] = None,
   @description("Git tag.")
-  git_tag: Option[String] = None
+  git_tag: Option[String] = None,
+
+  @description("The project config content used during build.")
+  project_config: Option[ViashProject] = None,
+  @description("The config mods passed in the CLI during build.")
+  config_mods: List[String] = Nil,
 )
