### 2023-12-13 a379de0
NextflowPlatform: Fix resolving relative paths wrt the param_list file (#592)
* Do not resolve remote paths relative to the --param_list file

* alternative implementation for resolving the sibling

* implement function using regex instead of nxf code

* Apply suggestions from code review

* check for nested paths

* fix helper code

* Add function tests

* Run tests only when nextflow installed

* Revert CHANGELOG changes

* move groovy tests to be run by scala

* update

* normalize path in nxf test

* try resolve issues between java 11 and higher with regex

* fix regex

* fix group

* more
[…]
