### 2023-12-14 089a365
NextflowPlatform: fix finding .build.yaml when worktree contains a symlink directory (#611)
* NextflowPlatform: fix finding .build.yaml when worktree contains symlink directory

* NextflowPlatform: fix finding .build.yaml when worktree contains symlink directory

* Add missing 'def'

* Add CHANGELOG entry

* Implement test
