+++ src/main/scala/io/viash/functionality/dependencies/Dependency.scala
@@ -167,7 +167,8 @@ object Dependency {
   }
 
   // Traverse the folder upwards until a `.build.yaml` is found but do not traverse beyond `repoPath`.
-  def findBuildYamlFile(path: Path, repoPath: Path): Option[Path] = {
+  def findBuildYamlFile(pathPossiblySymlink: Path, repoPath: Path): Option[Path] = {
+    val path = pathPossiblySymlink.toRealPath()
     val child = path.resolve(".build.yaml")
     if (Files.isDirectory(path) && Files.exists(child)) {
       Some(child)
