+++ CHANGELOG.md
@@ -6,6 +6,8 @@ TODO add summary
 
 `NextflowPlatform`: Do not resolve remote paths relative to the --param_list file (PR #592).
 
+`NextflowPlatform`: Allow finding `.build.yaml` file when worktree contains a directory that is a symlink (PR #611). 
+
 ## MINOR CHANGES
 
 * `NextflowTestHelper`: Do not hardcode a version of Nextflow in the testbench, 
