### 2023-12-14 efad6f1
Move project_config directly in the config instead of the info field
Add fix for copy pasta mistake in the dependency testbench
