### 2023-12-14 bad1139
Prepare release of Viash 0.8.2
