### 2023-12-15 5bc4214
Fix resolving nested workflows when the worktree contains symlinks.
