+++ CHANGELOG.md
@@ -5,7 +5,7 @@ TODO add summary
 ## BUG FIXES
 
 * `NextflowPlatform`: properly resolve paths when a nextflow workflow has another nextflow
-  workflow as dependency and the worktree contains a directory that is a symlink.
+  workflow as dependency and the worktree contains a directory that is a symlink (PR #614).
 
 # Viash 0.8.2 (2023-12-14): Minor changes and bug fixes
 
