### 2023-12-15 e55ac64
Fix resolving nested workflows when the worktree contains symlinks. (#614)
* Fix resolving nested workflows when the worktree contains symlinks.

* Add PR number to CHANGELOG

* Demonstrate failing test

* Revert "Demonstrate failing test"

This reverts commit 8ca5fc6b09f28bd9af2f0dafdde3c35ef4a38642.
