### 2023-12-22 viash-io/viash#617 @mberacochea
[BUG] Running "viash" with no options throws an error
### What happened?

I get the following error when running viash with no options

### Steps to reproduce

Download and install viash, and run `viash` with no options or flags

### Expected behavior

It should probably print the default help

### Relevant log output

```
$ viash
Unexpected error occurred! If you think this is a bug, please post
create an issue at https://github.com/viash-io/viash/issues containing
a reproducible example and the stack trace below.

viash - 0.7.1
Stacktrace:
java.util.NoSuchElementException: last of empty list
	at scala.collection.immutable.Nil$.last(List.scala:665)
	at scala.collection.immutable.Nil$.last(List.scala:661)
	at io.viash.Main$.mainCLI(Main.scala:178)
	at io.viash.Main$.mainCLIOrVersioned(Main.scala:117)
	at io.viash.Main$.main(Main.scala:64)
	at io.viash.Main.main(Main.scala)
```

### Version

OS: Ubuntu 22.04.3 LTS, Java: openjdk version "11.0.21" 2023-10-17, viash: 0.7.1

### Possible solution

_No response_

### Confirmation

- [X] I have searched the existing issues to make sure this is not a duplicate.
- [X] I have provided clear and concise information about the bug.

### Additional context

_No response_
