### 2023-12-22 viash-io/viash#617 @mberacochea
BTW, my viash installation is working
```
viash --help
viash 0.7.1 (c) 2020 Data Intuitive

viash is a spec and a tool for defining execution contexts and converting execution instructions to concrete instantiations.

This program comes with ABSOLUTELY NO WARRANTY. This is free software, and you are welcome to redistribute it under certain conditions. For more information, see our license at the link below.
  https://github.com/viash-io/viash/blob/master/LICENSE.md

Usage:
  viash run config.vsh.yaml -- [arguments for script]
  viash build config.vsh.yaml
  viash test config.vsh.yaml
  viash ns build
  viash ns test
  viash ns list
  viash config view config.vsh.yaml
  viash config inject config.vsh.yaml

Check the help of a subcommand for more information, or the API available at:
  https://viash.io

Arguments:
  -h, --help      Show help message
  -v, --version   Show version of this program

Subcommands:
  run
  build
  test
  ns
  config
```
