### 2023-12-22 viash-io/viash#617 @rcannood
Thanks @mberacochea. It appears this bug snuck in the release of Viash 0.6.4 as part of PR #294 :sweat_smile: I'm going to let the CI run to see whether #618 fixes the problem.
