### 2023-12-22 dd47691
add test for checking whether viash produces a helpful error message when no arguments are passed (#618)
* add test for checking whether viash produces a helpful error message when no arguments are passed

* Fix bug

* add entry to changelog
