+++ CHANGELOG.md
@@ -7,6 +7,9 @@ TODO add summary
 * `NextflowPlatform`: properly resolve paths when a nextflow workflow has another nextflow
   workflow as dependency and the worktree contains a directory that is a symlink (PR #614).
 
+* `Main`: Fixes a bug added by #294 which causes Viash to print a stacktrace instead of a helpful error message when `viash` is run without any arguments (#617, PR #618).
+  Thanks @mberacochea for pointing out this oversight!
+
 # Viash 0.8.2 (2023-12-14): Minor changes and bug fixes
 
 This release fixes a few bugs regarding dependencies and how the Nextflow platform handles Paths.
