### 2023-12-22 viash-io/viash#617 @rcannood
#618 has been merged into the 0.8.x develop branch and will be included in release 0.8.3.

Thanks @mberacochea for reporting this!
