### 2023-12-22 viash-io/viash#598 @rcannood
Contributing guidelines were added in #600 and merged into the main branch with the release of Viash 0.8.2. Thanks @abhishektiwari !
