### 2024-01-02 f8b0cdd
Set alias as key (#606)
* `Dependency`: When an alias is defined, pass the alias as a key to the `.run()` function.

* add test

* try to fix alias

* Add suggestions

Co-authored-by: Dries Schaumont <5946712+DriesSchaumont@users.noreply.github.com>

* fix test

---------

Co-authored-by: Dries Schaumont <5946712+DriesSchaumont@users.noreply.github.com>
