+++ src/test/scala/io/viash/platforms/nextflow/NextflowScriptTest.scala
@@ -135,6 +135,21 @@ class NextflowScriptTest extends AnyFunSuite with BeforeAndAfterAll {
     assert(exitCode == 0, s"\nexit code was $exitCode\nStd output:\n$stdOut\nStd error:\n$stdErr")
   }
 
+  test("Test whether aliasing works", DockerTest, NextflowTest) {
+    val (exitCode, stdOut, stdErr) = NextflowTestHelper.run(
+      mainScript = "target/nextflow/test_wfs/alias/main.nf",
+      args = List(
+        "--publish_dir", "output"
+      ),
+      cwd = tempFolFile
+    )
+
+    assert(exitCode == 0, s"\nexit code was $exitCode\nStd output:\n$stdOut\nStd error:\n$stdErr")
+    assert(stdOut.contains("base:step1_alias:proc"))
+    assert(stdOut.contains("base:step1:proc"))
+    assert(!stdOut.contains("Key for module 'step1' is duplicated"))
+  }
+
   test("Test for concurrency issues", DockerTest, NextflowTest) {
     val (exitCode, stdOut, stdErr) = NextflowTestHelper.run(
       mainScript = "target/nextflow/test_wfs/concurrency/main.nf",
