### 2024-01-05 #6089 @mberacochea
@rcannood Please see my feedback.

# General Comments

The paper is well-written and clearly explains the purpose and functionality of Viash. The authors provide detailed instructions on using the tool to describe modules with it. It is impressive in both the scope and the way it handles the whole process of pipeline writing, from initial module creation to fully functional pipelines using Nextflow.

# Documentation

The documentation provided is comprehensive and allowed me to successfully follow the tutorial to create a fully operational pipeline. It consists of a practical tutorial and extensive usage guidelines, as well as documentation on the tool's API. I encountered one minor issue (https://github.com/viash-io/viash/issues/617) which the authors promptly resolved.

# Source Code

While I am not familiar with Scala, I was able to understand the high-level logic by reading through the codebase, it is well structured and contains explanatory comments on complex sections.

## Generated Nextflow Code

The Nextflow code generated by Viash differs from the canonical style users may be accustomed to in other pipelines. This is not a flaw, but worth noting as it may cause initial confusion. Even though the produced modules are not meant to be modified, the ability to read and modify the code improves the developer experience, specially for the workflow developers. The paper could explore this a bit, as it’s mainly focus on modules and not so much on workflow development.

One question I have - could the custom Nextflow helper methods in Viash be abstracted/pushed into a Nextflow plugin or some kind of library? This question shouldn’t be considered a blocker for this PR.

# State of the field

My impression is that Viash provides functionality comparable to [nf-core/nf-tools](https://nf-co.re/tools) or [Snakemake Wrappers](https://snakemake-wrappers.readthedocs.io/en/stable/), though their approaches are different. Specially around the re-usable modules or components to build pipelines. Nf-core uses a different paradigm but it shares some of concepts with Viash. Explicitly contrasting nf-core/nf-tools - Snakemake Wrappers and Viash would benefit potenial users. The authors may also want to briefly mention [OpenPipelines](https://openpipelines.bio/) as a repository of Viash modules.

# Paper

## Applications in Bioinformatics

I don't think Viash is restricted solely to single cell analysis. Based on the underlying technologies, it seems widely applicable in bioinformatics (at least in omics).

In Figure 3, Tables B and C appear inadvertently merged. The authors may want to separate them for clarity
