+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Viash 0.x.x (yyyy-MM-dd): TODO Add title
+
+TODO add summary
+
 # Viash 0.8.3 (2024-01-08): Bug fixes
 
 Fix some edge cases and improve usability.
