+++ docs/paper/paper.md
@@ -63,7 +63,7 @@ Viash is available as an open-source project at [github.com/viash-io/viash](http
 
 
 # Statement of Need
-Recent developments in high-throughput RNA sequencing and imaging technologies allow present-day biologists to observe single-cell characteristics in ever more detail [@luecken_currentbestpractices_2019]. As the dataset size and the complexity of bioinformatics workflows increases, so does the need for scalable and reproducible data science. In single cell biology, recent efforts to standardise some of the most common single-cell analyses [@amezquita_orchestratingsinglecell_2020; @nfcoreframeworkcommunity_ewels2020; @huemos_bestpracticessingle_2023] tackle these challenges by using a workflow framework (e.g. Snakemake [@koster_snakemakescalablebioinformaticsworkflow_2012], Nextflow [@ditommaso_nextflowenablesreproducible_2017]), containerisation (e.g. Docker [@merkel_dockerlightweight_2014], Singularity/Apptainer [@kurtzer_singularityscientificcontainers_2017], Podman [@heon_podmantoolmanaging_2018]) and horizontal scaling in cloud computing (e.g. Kubernetes, HPC).
+Recent developments in high-throughput RNA sequencing and imaging technologies allow present-day biologists to observe single-cell characteristics in ever more detail [@luecken_currentbestpractices_2019]. As the dataset size and the complexity of bioinformatics workflows increases, so does the need for scalable and reproducible data science. In single cell biology, recent efforts to standardise some of the most common single-cell analyses [@amezquita_orchestratingsinglecell_2020; @nfcoreframeworkcommunity_ewels2020; @huemos_bestpracticessingle_2023] tackle these challenges by using a workflow framework (e.g. Snakemake, Nextflow), containerisation (e.g. Docker, Podman) and horizontal scaling in cloud computing (e.g. Kubernetes, HPC).
 
 Since research projects are increasingly more complex and interdisciplinary, researchers from different fields and backgrounds are required to join forces. This implies that not all project contributors can be experts in computer science. The chosen framework for such projects therefore needs to have a low barrier to entry in order for contributors to be able to participate. One common pitfall which greatly increases the barrier to entry is tightly coupling a workflow and the components it consists of. Major drawbacks include lower transparency of the overall workflow, limited reusability of workflow components, increased complexity, debugging time increase and a greater amount of time spent refactoring and maintaining boilerplate code. Non-expert developers in particular will experience more arduous debugging sessions as they need to treat the workflow as a black box.
 
@@ -92,20 +92,14 @@ One major benefit of using code regeneration is that best practices in workflow 
 
 
 # State of the field
-Nextflow, Snakemake, CWL, and WDL represent leading workflow management systems that the bioinformatics community increasingly relies upon.
 
-Nextflow stands out for its expertise in crafting data-driven computational pipelines, leveraging software containers to ensure both scalability and reproducibility, making it exceptionally suitable for demanding computational landscapes.
+The realm of bioinformatics workflow management is evolving rapidly, with numerous frameworks and portability solutions emerging to address the escalating complexity and scale of data processing [@wratten_reproduciblescalablesharable_2021]. Viash positions itself uniquely in this landscape as a meta-framework, focusing on the creation of portable workflow modules.
 
-Snakemake combines the intuitiveness of Python scripting with the reliability of a make-like execution methodology, providing both ease and efficiency.
[…]
