### 2024-01-10 f664897
update bibtex
