+++ docs/paper/library.bib
@@ -4,7 +4,7 @@
 
  @article{huemos_bestpracticessingle_2023, title={Best practices for single-cell analysis across modalities}, volume={24}, ISSN={1471-0064}, url={http://dx.doi.org/10.1038/s41576-023-00586-w}, DOI={10.1038/s41576-023-00586-w}, number={8}, journal={Nature Reviews Genetics}, publisher={Springer Science and Business Media LLC}, author={Heumos, Lukas and Schaar, Anna C. and Lance, Christopher and Litinetskaya, Anastasia and Drost, Felix and Zappia, Luke and Lücken, Malte D. and Strobl, Daniel C. and Henao, Juan and Curion, Fabiola and Aliee, Hananeh and Ansari, Meshal and Badia-i-Mompel, Pau and Büttner, Maren and Dann, Emma and Dimitrov, Daniel and Dony, Leander and Frishberg, Amit and He, Dongze and Hediyeh-zadeh, Soroor and Hetzel, Leon and Ibarra, Ignacio L. and Jones, Matthew G. and Lotfollahi, Mohammad and Martens, Laura D. and Müller, Christian L. and Nitzan, Mor and Ostner, Johannes and Palla, Giovanni and Patro, Rob and Piran, Zoe and Ramírez-Suástegui, Ciro and Saez-Rodriguez, Julio and Sarkar, Hirak and Schubert, Benjamin and Sikkema, Lisa and Srivastava, Avi and Tanevski, Jovan and Virshup, Isaac and Weiler, Philipp and Schiller, Herbert B. and Theis, Fabian J.}, year={2023}, month=mar, pages={550–572} }
 
- @article{luecken_currentbestpractices_2019, title={Current best practices in single‐cell RNA‐seq analysis: a tutorial}, volume={15}, ISSN={1744-4292}, url={http://dx.doi.org/10.15252/msb.20188746}, DOI={10.15252/msb.20188746}, number={6}, journal={Molecular Systems Biology}, publisher={EMBO}, author={Luecken, Malte D and Theis, Fabian J}, year={2019}, month=jun }
+ @article{luecken_currentbestpractices_2019, title={Current best practices in single‐cell RNA‐seq analysis: a tutorial}, volume={15}, ISSN={1744-4292}, url={http://dx.doi.org/10.15252/msb.20188746}, DOI={10.15252/msb.20188746}, number={6}, journal={Molecular Systems Biology}, publisher={Springer Science and Business Media LLC}, author={Luecken, Malte D and Theis, Fabian J}, year={2019}, month=jun }
 
  @article{nfcoreframeworkcommunity_ewels2020, title={The nf-core framework for community-curated bioinformatics pipelines}, volume={38}, ISSN={1546-1696}, url={http://dx.doi.org/10.1038/s41587-020-0439-x}, DOI={10.1038/s41587-020-0439-x}, number={3}, journal={Nature Biotechnology}, publisher={Springer Science and Business Media LLC}, author={Ewels, Philip A. and Peltzer, Alexander and Fillinger, Sven and Patel, Harshil and Alneberg, Johannes and Wilm, Andreas and Garcia, Maxime Ulysse and Di Tommaso, Paolo and Nahnsen, Sven}, year={2020}, month=feb, pages={276–278} }
 
@@ -19,6 +19,12 @@
   year = {2021},
   copyright = {Open Access}
 }
+
[…]
