+++ docs/paper/library.bib
@@ -56,8 +56,6 @@
   copyright = {Creative Commons Attribution 4.0 International}
 }
 
- @article{kurtzer_singularityscientificcontainers_2017, title={Singularity: Scientific containers for mobility of compute}, volume={12}, ISSN={1932-6203}, url={http://dx.doi.org/10.1371/journal.pone.0177459}, DOI={10.1371/journal.pone.0177459}, number={5}, journal={PLOS ONE}, publisher={Public Library of Science (PLoS)}, author={Kurtzer, Gregory M. and Sochat, Vanessa and Bauer, Michael W.}, editor={Gursoy, Attila}, year={2017}, month=may, pages={e0177459} }
-
 @article{crusoe_methodsincludestandardizing_2022,
   title={Methods included: standardizing computational reuse and portability with the Common Workflow Language},
   volume={65},
