### 2024-01-10 #6089 @rcannood
Hey all! I went through the comments I received so far and tried to address each of the reviewers' concerns. Please let me know if you have any further feedback.

# @abhishektiwari's comments

## Community Guidelines

> Please add https://github.com/viash-io/viash/issues/598. I also think project will benefit from issue and pull request templates.

Done, done, and done! Thanks for the suggestion!

## Title consistency

> For consistency, I would suggest paper headline to use data pipeline modules and not workflow modules. Everywhere else in the paper and documentation website authors reference Viash as solution for modular data pipelines. Having said that data pipelines are a special type of workflows with major focus on ETL, data cleaning, integration, and preparation for analysis or visualization, so I am also fine using workflow.

Good point! For the sake of consistency, I ended up replacing all instances of "pipeline" with "workflow" in https://github.com/viash-io/viash/commit/5029fe049a981bf37816354e181fc26757161fd2

## State of the field

> Current state of the field can be further improved. Given the focus on Bioinformatics related computation/data pipeline frameworks, Viash, and it's counterparts Nextflow, Snakemake, CWL belong to Pipeline as code category. Please consider including [Toil](http://toil.ucsc-cgl.org/) in the list as it supports script based pipeline definitions but also supports CWL and WDL. Plus there is other category of frameworks such as Galaxy and KNIME Workbench used by bioinformatics community those rely on UI based drag-n-drop pipeline builders – something worth mentioning here. Similarly, more generic Pipeline as code frameworks such Apache Airflow and Luigi can use standalone executable generated by Viash and used by bioinformatics community.

Thanks for this suggestion! I completely rewrote the state of the field in https://github.com/viash-io/viash/commit/6a21f69a22add86bbf5f7f217e367e1a434ad9ac . Rather than only compare Viash to Snakemake, Nextflow, CWL and WDL, I instead compare Viash to workflow managers and portability solutions. Based on https://doi.org/10.1038/s41592-021-01254-9, we then discuss the different types of workflow managers (graphical, programmatic, specification-based) and portability solutions (package manager, containerization).

## Suggested improvements (non-blocking)

> Given the Viash has broader applicability beyond Bioinformatics particularly MLOps Pipelines, I suggest adding starter using other more generic frameworks such Airflow. In case of Apache Airflow, Airflow [BashOperator](https://airflow.apache.org/docs/apache-airflow/stable/howto/operator/bash.html) will works just fine with standalone executable generated by Viash.

Thanks for your suggestion,  we will definitely take this into account when extending our documentation with different workflow frameworks!

# @mberacochea's comments

## Generated Nextflow Code

> One question I have - could the custom Nextflow helper methods in Viash be abstracted/pushed into a Nextflow plugin or some kind of library? This question shouldn’t be considered a blocker for this PR.

I would really like that! The latest release of Viash already split up the helper functions into separate helper files per function so we can unit test them separately. I would like some of this functionality to be included in a Groovy library so we don't need to include it in every Nextflow module that is generated.

## State of the field

> My impression is that Viash provides functionality comparable to [nf-core/nf-tools](https://nf-co.re/tools) or [Snakemake Wrappers](https://snakemake-wrappers.readthedocs.io/en/stable/), though their approaches are different. Specially around the re-usable modules or components to build pipelines. Nf-core uses a different paradigm but it shares some of concepts with Viash. Explicitly contrasting nf-core/nf-tools - Snakemake Wrappers and Viash would benefit potenial users. The authors may also want to briefly mention [OpenPipelines](https://openpipelines.bio/) as a repository of Viash modules.

Thanks for making the connection with nf-core tools (I think you mean nf-core/modules instead?), Snakemake wrappers and OpenPipelines. I did not include references to these projects in the paper as I felt that creating a repository with reusable commonly used bioinformatics tools goes way beyond the scope of this manuscript.

## Applications in bioinformatics

> I don't think Viash is restricted solely to single cell analysis. Based on the underlying technologies, it seems widely applicable in bioinformatics (at least in omics).

Good point. I dropped the word "single-cell" in https://github.com/viash-io/viash/commit/874169afc30b620aee05baf237741bf5d9f4711e

## Paper figures

> In Figure 3, Tables B and C appear inadvertently merged. The authors may want to separate them for clarity

Thanks, I fixed this in https://github.com/viash-io/viash/commit/758db0f37cb23e2d01e321dd7dcfd8cdd69fb670 !
