### 2024-01-15 c25bc58
Fix dependency resolving when symlinks are used in paths
