+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Viash 0.x.x (yyyy-MM-dd): TODO Add title
+
+TODO add summary
+
 # Viash 0.8.4 (2024-01-15): Bug fix
 
 Fix building components with dependencies that have symlinks in their paths.
