### 2024-01-15 ad92655
Add extra fields, no supporting logic yet
