### 2024-01-17 2459439
Add fields in the component config and set some values from the project config
Refactor some other code to use lenses a bit more
