### 2024-01-17 e332c9d
Add changelog entry
