### 2024-01-17 ad0a22d
Add helper method to anonymize paths and add tests
