### 2024-01-17 11ab8ff
If the path can't be relativized, go for the drastic solution to fully anonymize the path except for the last path section
