### 2024-01-17 #6089 @wilkinson
@danielskatz I finished re-reading after the edits, and I am done with my review. The latest changes from @rcannood do help, and I can only really think of one (non-blocking) critique: in Figure 1, the box labeled "Script" just looks kind of empty and the linebreaks are kind of strange. I'm not sure how I would suggest to edit that, and it does not really matter. Overall, I am very impressed with this tool, the paper, and the authors!
