### 2024-01-18 3d0c92a
Experimental changes to reorganize the location of project dir
