+++ src/main/scala/io/viash/Main.scala
@@ -168,7 +168,7 @@ object Main extends Logging {
     */
   def mainCLI(args: Array[String], workingDir: Option[Path] = None): Int = {
     // try to find project settings
-    val proj0 = workingDir.map(ViashProject.findViashProject(_)).getOrElse(ViashProject())
+    val (proj0, projectDir) = workingDir.map(ViashProject.findViashProject(_)).getOrElse((ViashProject(), None))
 
     // strip arguments meant for viash run
     val (viashArgs, runArgs) = {
@@ -223,7 +223,11 @@ object Main extends Logging {
     // process commands
     cli.subcommands match {
       case List(cli.run) =>
-        val config = readConfig(cli.run, project = proj1)
+        val config = readConfig(
+          cli.run,
+          project = proj1,
+          projectDir = projectDir
+        )
         ViashRun(
           appliedConfig = config,
           args = runArgs.toIndexedSeq.dropWhile(_ == "--"), 
@@ -232,8 +236,12 @@ object Main extends Logging {
           memory = cli.run.memory.toOption
         )
       case List(cli.build) =>
-        val config = readConfig(cli.build, project = proj1)
-        val config2 = singleConfigDependencies(config, cli.build.output.toOption, proj1.rootDir)
+        val config = readConfig(
+          cli.build,
+          project = proj1,
+          projectDir = projectDir
+        )
+        val config2 = singleConfigDependencies(config, cli.build.output.toOption)
         val buildResult = ViashBuild(
           appliedConfig = config2,
           output = cli.build.output(),
@@ -242,8 +250,12 @@ object Main extends Logging {
         )
         if (buildResult.isError) 1 else 0
       case List(cli.test) =>
-        val config = readConfig(cli.test, project = proj1)
-        val config2 = singleConfigDependencies(config, None, proj1.rootDir)
+        val config = readConfig(
+          cli.test,
+          project = proj1,
+          projectDir = projectDir
+        )
+        val config2 = singleConfigDependencies(config, None)
         ViashTest(
           config2,
           keepFiles = cli.test.keep.toOption.map(_.toBoolean),
@@ -253,8 +265,12 @@ object Main extends Logging {
         )
         0 // Exceptions are thrown when a test fails, so then the '0' is not returned but a '1'. Can be improved further.
       case List(cli.namespace, cli.namespace.build) =>
-        val configs = readConfigs(cli.namespace.build, project = proj1)
-        val configs2 = namespaceDependencies(configs, proj1.target, proj1.rootDir)
+        val configs = readConfigs(
+          cli.namespace.build,
+          project = proj1,
+          projectDir = projectDir
+        )
+        val configs2 = namespaceDependencies(configs, proj1.target)
         var buildResults = ViashNamespace.build(
           configs = configs2,
           target = proj1.target.get,
@@ -268,9 +284,13 @@ object Main extends Logging {
           .count(_.isError)
         if (errors > 0) 1 else 0
       case List(cli.namespace, cli.namespace.test) =>
-        val configs = readConfigs(cli.namespace.test, project = proj1)
+        val configs = readConfigs(
+          cli.namespace.test,
+          project = proj1,
+          projectDir = projectDir
+        )
         // resolve dependencies
-        val configs2 = namespaceDependencies(configs, None, proj1.rootDir)
+        val configs2 = namespaceDependencies(configs, None)
         // flatten engines
         val configs3 = configs2.flatMap{ ac => 
           ac.engines.map{ engine => 
@@ -296,11 +316,12 @@ object Main extends Logging {
         val configs = readConfigs(
           cli.namespace.list,
           project = proj1,
[…]
