### 2024-01-18 f815bc3
Fix testbench and clean up code a bit
