### 2024-01-18 #6089 @mberacochea
@rcannood great stuff, looking forward to see how viash keeps evolving.

@danielskatz My review is completed, my recommendation is to accept and publish this one.
