### 2024-01-18 #6089 @rcannood
Just updated figure 1 based on @wilkinson's feedback!
