### 2024-01-22 5883985
Only set component license and organization if it was unset
Forgot to add the checks initially
