### 2024-01-22 457159d
Reduce & simplify lens code
