### 2024-01-22 c246525
Fix config mods AND project config repositories together throwing exceptions
.functionality.repositories was still outputting internalFunctionality fields
Use wrapper method to parse config after applying config mods, it gives much better information if something goes wrong
