+++ docs/paper/paper.md
@@ -63,26 +63,26 @@ Viash is available as an open-source project at [github.com/viash-io/viash](http
 
 
 # Statement of Need
-Recent developments in high-throughput RNA sequencing and imaging technologies allow present-day biologists to observe single-cell characteristics in ever more detail [@luecken_currentbestpractices_2019]. As the dataset size and the complexity of bioinformatics workflows increases, so does the need for scalable and reproducible data science. In single cell biology, recent efforts to standardise some of the most common single-cell analyses [@amezquita_orchestratingsinglecell_2020; @nfcoreframeworkcommunity_ewels2020; @huemos_bestpracticessingle_2023] tackle these challenges by using a workflow framework (e.g. Snakemake, Nextflow), containerisation (e.g. Docker, Podman) and horizontal scaling in cloud computing (e.g. Kubernetes, HPC).
+Recent developments in high-throughput RNA sequencing and imaging technologies allow present-day biologists to observe single-cell characteristics in ever more detail [@luecken_currentbestpractices_2019]. As the dataset size and the complexity of bioinformatics workflows increases, so does the need for scalable and reproducible data science. In single cell biology, recent efforts to standardise some of the most common single-cell analyses [@amezquita_orchestratingsinglecell_2020; @nfcoreframeworkcommunity_ewels2020; @huemos_bestpracticessingle_2023] tackle these challenges by using a workflow framework (e.g., Snakemake, Nextflow), containerisation (e.g., Docker, Podman), and horizontal scaling in cloud computing (e.g., Kubernetes, HPC).
 
-Since research projects are increasingly more complex and interdisciplinary, researchers from different fields and backgrounds are required to join forces. This implies that not all project contributors can be experts in computer science. The chosen framework for such projects therefore needs to have a low barrier to entry in order for contributors to be able to participate. One common pitfall which greatly increases the barrier to entry is tightly coupling a workflow and the components it consists of. Major drawbacks include lower transparency of the overall workflow, limited reusability of workflow components, increased complexity, debugging time increase and a greater amount of time spent refactoring and maintaining boilerplate code. Non-expert developers in particular will experience more arduous debugging sessions as they need to treat the workflow as a black box.
+Since research projects are increasingly more complex and interdisciplinary, researchers from different fields and backgrounds are required to join forces. This implies that not all project contributors can be experts in computer science. The chosen framework for such projects therefore needs to have a low barrier to entry in order for contributors to be able to participate. One common pitfall that greatly increases the barrier to entry is tightly coupling a workflow and the components it consists of. Major drawbacks include lower transparency of the overall workflow, limited reusability of workflow components, increased complexity, increased debugging time, and a greater amount of time spent refactoring and maintaining boilerplate code. Non-expert developers in particular will experience more arduous debugging sessions as they need to treat the workflow as a black box.
 
-In this work we introduce Viash, a tool for speeding up workflow prototyping through code generation, component modularity and separation of concerns. With Viash, a user can create a workflow module by writing a small script or using a pre-existing code block, adding a small amount of metadata, and using Viash to generate the boilerplate code needed to turn it into a modular Nextflow component. This separates the component functionality from the workflow workflow, thereby allowing a component developer to focus on implementing the required functionality using the domain-specific toolkit at hand while being completely agnostic to the chosen workflow framework. Similarly, a workflow developer can design a workflow by chaining together Viash modules while being completely agnostic to the scripting language used in the component.
[…]
