### 2024-01-23 #6089 @danielskatz
@rcannood - I've suggested some edits in https://github.com/viash-io/viash/pull/628.

Also, I'm not sure if "This separates the component functionality from the workflow workflow" is correct or not. Do you want to say "workflow workflow"?

Once you've merged the PR and perhaps changed the additional line I mention, please
