### 2024-01-23 316cf35
Add the reference case class as described/requested in the PR
