### 2024-01-23 2a030b6
improve the documentation for references and links
