+++ src/main/scala/io/viash/project/ProjectConfigReferences.scala
@@ -5,7 +5,7 @@ import io.viash.helpers.data_structures.OneOrMore
 
 @description("References to external resources related to the project.")
 @since("Viash 0.9.0")
-case class ViashProjectReferences(
+case class ProjectConfigReferences(
   @description("One or multiple DOI reference(s) of the project.")
   @example("doi: 10.1000/xx.123456.789", "yaml")
   doi: OneOrMore[String] = Nil,
