### 2024-01-24 cbf1c29
Add testbench to verify copying project config values to component config
Does not test dependency repositories
