### 2024-01-24 dc89e7e
Add test to check project config repositories are copied to .functionality.repositories
pretty basic test, doesn't check order etc (which should be irrelevant anyway)
