+++ src/main/scala/io/viash/lenses/ConfigLenses.scala
@@ -27,9 +27,6 @@ import FunctionalityLenses._
 
 object ConfigLenses {
   val functionalityLens = GenLens[Config](_.functionality)
-  val keywordsLens = GenLens[Config](_.keywords)
-  val licenseLens = GenLens[Config](_.license)
-  val organizationLens = GenLens[Config](_.organization)
   val enginesLens = GenLens[Config](_.engines)
   val runnersLens = GenLens[Config](_.runners)
 
@@ -40,4 +37,7 @@ object ConfigLenses {
   val composedTestResourcesLens = functionalityLens ^|-> testResourcesLens
   val composedDependenciesLens = functionalityLens ^|-> dependenciesLens
   val composedRepositoriesLens = functionalityLens ^|-> repositoriesLens
+  val composedKeywordsLens = functionalityLens ^|-> keywordsLens
+  val composedLicenseLens = functionalityLens ^|-> licenseLens
+  val composedOrganizationLens = functionalityLens ^|-> organizationLens
 }
