+++ docs/paper/paper.md
@@ -67,7 +67,7 @@ Recent developments in high-throughput RNA sequencing and imaging technologies a
 
 Since research projects are increasingly more complex and interdisciplinary, researchers from different fields and backgrounds are required to join forces. This implies that not all project contributors can be experts in computer science. The chosen framework for such projects therefore needs to have a low barrier to entry in order for contributors to be able to participate. One common pitfall that greatly increases the barrier to entry is tightly coupling a workflow and the components it consists of. Major drawbacks include lower transparency of the overall workflow, limited reusability of workflow components, increased complexity, increased debugging time, and a greater amount of time spent refactoring and maintaining boilerplate code. Non-expert developers in particular will experience more arduous debugging sessions as they need to treat the workflow as a black box.
 
-In this work we introduce Viash, a tool for speeding up workflow prototyping through code generation, component modularity, and separation of concerns. With Viash, a user can create a workflow module by writing a small script or using a pre-existing code block, add a small amount of metadata, and use Viash to generate the boilerplate code needed to turn it into a modular Nextflow component. This separates the component functionality from the workflow workflow, thereby allowing a component developer to focus on implementing the required functionality using the domain-specific toolkit at hand while being completely agnostic to the chosen workflow framework. Similarly, a workflow developer can design a workflow by chaining together Viash modules while being completely agnostic to the scripting language used in the component.
+In this work we introduce Viash, a tool for speeding up workflow prototyping through code generation, component modularity, and separation of concerns. With Viash, a user can create a workflow module by writing a small script or using a pre-existing code block, add a small amount of metadata, and use Viash to generate the boilerplate code needed to turn it into a modular Nextflow component. This separates the component functionality from the workflow logic, thereby allowing a component developer to focus on implementing the required functionality using the domain-specific toolkit at hand while being completely agnostic to the chosen workflow framework. Similarly, a workflow developer can design a workflow by chaining together Viash modules while being completely agnostic to the scripting language used in the component.
 
 
 
