### 2024-01-24 #6089 @rcannood
* Thanks for your improvements! I merged your PR.
* Thanks for spotting this refactoring error! I changed "workflow workflow" into "workflow logic".
* Finally, the software in Zenodo was archived at https://zenodo.org/records/10563479, doi: [10.5281/zenodo.10563479](https://doi.org/10.5281/zenodo.10563479)
