### 2024-01-24 #6089 @rcannood
A big thank you to @abhishektiwari @mberacochea @wilkinson and @danielskatz !

I see the names of Eric and Dries are split up incorrectly. Should I create a PR to amend https://github.com/openjournals/joss-papers/pull/4948?
