### 2024-01-30 #6089 @danielskatz
@rcannood - The files you are trying to change in your PR are automatically generated by JOSS. So please delete this PR, and instead, make changes in your paper source.  See https://joss.readthedocs.io/en/latest/submitting.html#names in the JOSS documentation for how to do this.

Let me know when this is done, and I'll redo the generation, so we can check the xml and jats files, then when we are happy, I'll republish those versions.
