### 2024-02-11 #6089 @rcannood
Hi @danielskatz! Thanks for your help! I had missed your reply until Arfon pointed it out to me.

I made the necessary changes to the author names in the `paper` branch at `viash-io/viash`. Would you be able to regenerate the XML and Jats files?

Thanks you so much for your time! :bow:
